requirements.txt
praw
python-dotenv
tweepy
requests
